# [Enhancement] Enhance NormalizePredicateRule for many Or/And