--- a/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rewrite/ScalarOperatorRewriter.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rewrite/ScalarOperatorRewriter.java
@@ -29,8 +29,6 @@ import com.starrocks.sql.optimizer.rewrite.scalar.ReduceCastRule;
 import com.starrocks.sql.optimizer.rewrite.scalar.ScalarOperatorRewriteRule;
 import com.starrocks.sql.optimizer.rewrite.scalar.SimplifiedPredicateRule;
 import com.starrocks.sql.optimizer.rewrite.scalar.SimplifiedScanColumnRule;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 import java.util.List;
 
@@ -72,7 +70,6 @@ public class ScalarOperatorRewriter {
             new ExtractCommonPredicateRule(),
             new ArithmeticCommutativeRule()
     );
-    private static final Logger LOG = LogManager.getLogger(ScalarOperatorRewriter.class);
     private final ScalarOperatorRewriteContext context;
 
     public ScalarOperatorRewriter() {
@@ -87,11 +84,7 @@ public class ScalarOperatorRewriter {
         do {
             changeNums = context.changeNum();
             for (ScalarOperatorRewriteRule rule : ruleList) {
-                long begin = System.currentTimeMillis();
                 result = rewriteByRule(result, rule);
-                long end = System.currentTimeMillis();
-                long duration = end - begin;
-                LOG.info("rule:" + rule + " time:" + duration);
             }
 
             if (changeNums > Config.max_planner_scalar_rewrite_num) {