Why I'm doing:
Expr like 100 or takes too long in transform phase, Especially NormalizePredicateRule‘s visitCompoundPredicate, because its time complexity is O(n^2) n is tree node numbers.

What I'm doing:
when visit expr tree bottom up, if child is compound operator, cache child's unique compound tree leaf node in set and child's hash value to eliminate duplicate calculations

before this pr:
```
+-------------------------------------------------------+
| Explain String                                        |
+-------------------------------------------------------+
|   37ms|-- Total[1] 4s479ms                            |
|   37ms|    -- Analyzer[1] 5ms                         |
|   44ms|    -- Transformer[1] 2s892ms                  |
| 2936ms|    -- Optimizer[1] 990ms                      |
| 2936ms|        -- preprocessMvs[1] 0                  |
| 2936ms|        -- RuleBaseOptimize[1] 967ms           |
| 3906ms|        -- CostBaseOptimize[1] 18ms            |
| 3924ms|        -- PhysicalRewrite[1] 0                |
| 3925ms|        -- PlanValidate[1] 1ms                 |
| 3925ms|            -- InputDependenciesChecker[1] 1ms |
| 3926ms|            -- TypeChecker[1] 0                |
| 3926ms|            -- CTEUniqueChecker[1] 0           |
| 3926ms|            -- ColumnReuseChecker[1] 0         |
| 3926ms|    -- ExecPlanBuild[1] 589ms                  |
| Tracer Cost: 63us                                     |
+-------------------------------------------------------+
15 rows in set (4.52 sec)
```

after:
```
+-------------------------------------------------------+
| Explain String                                        |
+-------------------------------------------------------+
|   41ms|-- Total[1] 2s297ms                            |
|   41ms|    -- Analyzer[1] 6ms                         |
|   49ms|    -- Transformer[1] 1s209ms                  |
| 1258ms|    -- Optimizer[1] 273ms                      |
| 1258ms|        -- preprocessMvs[1] 0                  |
| 1258ms|        -- RuleBaseOptimize[1] 236ms           |
| 1497ms|        -- CostBaseOptimize[1] 31ms            |
| 1529ms|        -- PhysicalRewrite[1] 0                |
| 1529ms|        -- PlanValidate[1] 1ms                 |
| 1529ms|            -- InputDependenciesChecker[1] 1ms |
| 1531ms|            -- TypeChecker[1] 0                |
| 1531ms|            -- CTEUniqueChecker[1] 0           |
| 1531ms|            -- ColumnReuseChecker[1] 0         |
| 1531ms|    -- ExecPlanBuild[1] 807ms                  |
| Tracer Cost: 63us                                     |
+-------------------------------------------------------+
15 rows in set (2.35 sec)
```


Fixes #issue

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:

- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5